requirements.txt
asgiref==3.8.1
certifi==2024.7.4
channels==4.1.0
charset-normalizer==3.3.2
cloudinary==1.40.0
Django==5.0.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
gunicorn==22.0.0
idna==3.7
packaging==24.1
pillow==10.3.0
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dotenv==1.0.1
pytz==2024.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2